Kommentar update

+++ SWC/Aktiv.java
@@ -3,8 +3,8 @@ import java.io.*;
 
 public class Aktiv extends Kontingent {
    public Aktiv(double pris) {
-      super((int)pris);
+      super(pris);
    } 
    
-   public double getPris() { return (int)super.getPris(); }
+   public double getPris() { return super.getPris(); }
 }
